requirements.txt
requests>=2.31.0
pyyaml>=6.0
aiohttp>=3.9.0
pytest>=7.4.0
pytest-cov>=4.1.0
